fix: initialize secret handler lazily
The secret handler needs access to the secret store. This store is usually
depending on environment-specific data provided when creating the specific
context. Creating the secret handler in the default constructor can easily
fail since the environment-specific data is not there yet (the super
constructor is called before constructing the context implementation
and setting its specific data). Thus, we need to create it transparently
at a later point in time.