feat(github): invoke workflow (create workflow dispatch event) from element template (#1634)
* chore: increment Github connector version

To prepare for incoming new options

* feat: Add Actions operation group

* feat: Add workflow dispatch operation type

* feat: Display repository owner

* feat: Display Repository field

* feat: Display Workflow ID field

* feat: Display Git reference field

* feat: Display workflow inputs field

* fix: Shorten Git reference description

To make more consistent with other descriptions and reduce unnecessary
info

* fix: clarify workflow input format

Use more concise language than in the Github documentation

* fix: add file name as possible workflow ID

As documented in the Github API. In my experience so far, file name is
much easier to find than workflow ID.

* fix: remove default op type

There appears to be a bug in Desktop Modeler.

When there is only one choice in the dropdown, and a default is applied,
 other fields dependent on that choice will never bind.

 This even happens when selecting another operation group, then
 re-selecting the operation group.

* fix: add URL variable for workflows

* fix: add optional flag for workflow inputs

Without it, Modeller will fail to deploy

* fix: add POST method to create workflow

* fix: add missing forward slash to URL

* feat: add request body for workflow dispatch

* fix: make workflowInput follow JSON convention

As done in Google Drive and WhatsApp connectors.

Also, adjust inputs to work for non-null input. Will need to test with
other null input tomorrow.

* chore: move workflow objects further down

To not interfere visually with fields for other operation groups

* chore: remove "workflowInputs" id

Inconsistent with other fields in this connector.

We don't yet need to reference workflowInputs in a condition

* feat: add default {} to workflowInputs

Because a single JSON object is needed for this field.

This is similar to the approach used in WhatsApp

* chore: move inputs to correct place

Erroneously put them with body objects before

* fix: remove reference to JSON

Part of QA changes

"JSON object" was confusing, because valid FEEL objects will be
converted to JSON before being sent to GitHub

* fix: use GitHub API doc for Workflow ID

Part of QA changes

* fix: Align description with entry in camunda-docs