fix: revert to v4 EasyPost template
v5 was never released to customers.

v5 will not work with runtimes before 8.5.0.

We don't yet have a
mechanism to inform customers which runtimes a template is supported by.

Because this template was only cleaning up technical debt and not
introducing new features, we believe the best option is to not release a
 new version for now.