Update RELEASE.yaml -  add option to checkout tag if it already exists
target_commitish only exists when tag doesn't exist already when the release is created.

While this is often useful, there are times when the tag already exists.

For example, if someone edits an existing release instead of creates a new one. Or if the GitHub release needs to be re-created for some reason.