ci(release): support existing tag, clarify Maven staging (#2451)
* Update RELEASE.yaml -  add option to checkout tag if it already exists

target_commitish only exists when tag doesn't exist already when the release is created.

While this is often useful, there are times when the tag already exists.

For example, if someone edits an existing release instead of creates a new one. Or if the GitHub release needs to be re-created for some reason.

* Update RELEASE.yaml - use tag name first

For our purposes, `tag_name` is already populated by the time the workflow runs - even when the tag doesn't already exist once the release object is created.

* Update RELEASE.yaml - clarify deploy to Maven Central staging

* Update RELEASE.yaml - Add Helm chart integration tests

To catch any potential issues introduced in the feature branch

* Update RELEASE.yaml - add default section

As done here: https://github.com/camunda/connectors/pull/2059/files

* Update RELEASE.yaml - use existing Helm integration tests

* WIP: enable upgrade flow

Trying to use older charts for older minor version patches

* WIP: remove upgrade flows

behaviour wasn't as expected

* WIP: overwrite non-connectors with 8.3.10

Seeing if this will work with testing v8.3.11

* WIP: copy values from values-v8.3.yaml

To try to replicate SM setup

* WIP: use camunda-helm-git-ref

* WIP: add global image tag

Although integration test run was successful without it, I want to try
with it as well

* ci: remove Helm int tests from this branch

Since the change will be more complicated, I'll create a new PR for it.