ci(integration): prepare integration test for use in release and deploy snapshots workflows (#2666)
* ci(integration): Use integration test workflow in release and deploy snapshots

* Update DEPLOY_SNAPSHOTS.yaml - re-use existing integration test workflow

* Update RELEASE.yaml - run as separate job

As is done in update snapshots workflow

* Update DEPLOY_SNAPSHOTS.yaml - inherit secrets

* Update INTEGRATION_TEST.yml - add on workflow_call, remove on push

- Add on workflow_call to support calling from other workflows (e.g., deploy snapshot, release
- Remove on push because it appears we can only run one integration test total at a time. Renovate often "pushes" many commits at once. In addition, integration tests are too flaky to be actioned on each time.

* Update DEPLOY_SNAPSHOTS.yaml - pass in main as helm branch

* Update helm-git-refs.json - use short branch names

For easier use with other workflows. integration test workflow will follow.

* Update INTEGRATION_TEST.yml - use short branch name

So we can more easily use either passed-in arg or where integration test is being run.

Hardcoded .json file already edited

* Update RELEASE.yaml - pass in release branch

* Update DEPLOY_SNAPSHOTS.yaml - pass in release-branch

External workflows shouldn't need to use helm-branch

* ci: revert snapshot and release changes

For workflows in the same repo, we can only call the workflow on the
main branch.

Because of this I'd like to merge in the integration test changes alone
first.

That way, we can test deploy and release before merging