Create reopen_support_issues_without_version_label.yml
Adding a GHA check on issues that are closed that if it includes a "support" label. If it does, this script will check that if it has the version: -label included. If it does not, then the script will reopen the issue and comment on it that this label needs to be added before closure.