refactor(agentic-ai): Align `fromAi()` parameter extraction with core engine logic (#5205)
* refactor(agentic-ai): restrict toolCall parameters in schema generation to be prefixed with toolCall.

* refactor(agentic-ai): allow excluding a property from being required by setting the required option

* refactor(agentic-ai): return full reference path from fromAi() parameter extraction

This is to be aligned with the core engine implementation. The schema
generation will take care of only using parameter within the defined
namespace (toolCall.* by default).

* refactor(agentic-ai): align fromAi() parameter extraction with core engine implementation

Primarily moves logic into the same places as in the core engine
implementation, but changes:

Arguments now need to be constant values instead of expressions. This
avoids confusion while modeling as the fromAi() function have another
evaluation context as normal expressions (hence no variables are
available).

* refactor(agentic-ai): use the zeebe-tagged-parameters module to handle parameter extraction