fix(agentic-ai): only set ResponseFormat if it is configured to be JSON, keep it null/default otherwise
The Anthropic chat model throws a validation exception as soon as the response format is set (even to TEXT)
as it only supports text responses. While this is  a bug which should be addressed upstream, we can work around
this issue by only configuring the format if we explicitely want to use JSON mode.